Allow to generate avatars from any string
FossilOrigin-Name: fc59eb1f3bf107bc7a52c69f6743a3aa694e3467fcf9193b38c94a1bd9885167

+++ src/include/lib/Paheko/Web/Router.php
@@ -93,7 +93,7 @@ class Router
 		}
 		// Users avatars
 		elseif ($first === 'user' && strpos($uri, 'user/avatar/') !== false) {
-			Users::serveAvatar((int)substr($uri, strlen('user/avatar/')));
+			Users::serveAvatar(substr($uri, strlen('user/avatar/')));
 			return;
 		}
 		// Add trailing slash to URLs if required
